Transform HomePage component to App
